Rule of six date
I'll be honest, I don't remember the timeline as it happened, but I was looking at these intervention dates and the rule of six appears to have come into force on 14th Sep, not 9th as in the code (source: https://www.bbc.co.uk/newsround/54088842). If that's not what you're tracking here, feel free to close this.