genomics/lineage: fix to plotting in case of multi-parent lineages
Some lineages are the result of recombination, and have multiple
parents. Previously that case wasn't handled and the plotting errors
out. Here pick for the parent of such a lineage the lineage with
the most dots in the name (ie. longest chain, most mutations, newest?)

Does not handle ties (when the parents have the same length chain),
though such lineages do not exists as of yet.